requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Streamlit for Web UI
streamlit>=1.31.0

# RAG components - Embedding & Vector Store
sentence-transformers>=2.3.0
faiss-cpu>=1.8.0
numpy>=1.26.0
rank_bm25>=0.2.2

# LLM - OpenAI only
openai>=1.12.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
